Move stuff to plugins
