+++ twake/frontend/src/app/components/RichTextEditor/EditorDataParser.ts
@@ -7,11 +7,13 @@ export function toString(editorState: EditorState, format: EditorTextFormat): st
   let contentState = editorState.getCurrentContent();
   switch (format) {
     case "markdown": {
-      const markdown: string = stateToMarkdown(contentState, { gfm: true }).trim();
+      let markdown: string = stateToMarkdown(contentState, { gfm: true }).trim();
       // when empty the stateToMarkdown lib returns a zero-width-space character in first position
       if (markdown.length === 1 && markdown.charCodeAt(0) === 8203) {
-        return "";
+        markdown = "";
       }
+
+      console.log("MARKDOWN", `${markdown}`)
 
       return markdown;
     }
